requirements.txt
cachetools==3.1.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==1.11.18
django-filebrowser==3.11.1
django-grappelli==2.12.2
django-qrcode==0.3
djangorestframework==3.9.1
docopt==0.4.0
futures==3.3.0
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.8
kml2geojson==4.0.2
oauthlib==3.0.2
Pillow==6.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.18.2
PyJWT==1.7.1
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.5
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
uritemplate==3.0.1
urllib3==1.24.1
zpl==0.1.1
